refactor: move jwt auth related code to shared context
